DEB: disable service on purge instead of removal, restart on upgrade
systemctl refuses to disable inexistent units. deb-systemd-helper can do the job for us, preserving
enablement state on removal.

try-restart on upgrade is problematic because if the package is removed without purge, the next install is
considered an upgrade as the version is still passed as postinst's second command line argument, leading to a
stopped service. Let's just restart unconditionally.